# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IIoT Platform API
  version: 0.2.0
  description: |
    API do backend da plataforma IIoT.
    Esta especificacao documenta endpoints atuais (auth, devices, telemetry).
servers:
  - url: http://localhost:3001
tags:
  - name: Health
  - name: Auth
  - name: Devices
  - name: Telemetry

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API_KEY
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        request_id:
          type: string
      required: [error]
    UserObject:
      type: object
      properties:
        user_id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid, nullable: true }
        email: { type: string }
        role: { type: string, enum: [super_admin, tenant_admin, tenant_user] }
        status: { type: string, enum: [active, suspended, deleted] }
        created_at: { type: string, format: date-time }
        last_login_at: { type: string, format: date-time, nullable: true }
    AuthResponse:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        expires_in: { type: integer, example: 3600 }
        user: { $ref: "#/components/schemas/UserObject" }
    RefreshResponse:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        expires_in: { type: integer, example: 3600 }
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8, example: "Abcdef1!" }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    ClaimRequest:
      type: object
      required: [device_id, claim_code]
      properties:
        device_id: { type: string, format: uuid }
        claim_code: { type: string }
    ClaimResponse:
      type: object
      properties:
        device_id: { type: string, format: uuid }
        message: { type: string }
    ProvisionRequest:
      type: object
      properties:
        device_label: { type: string, maxLength: 50, example: "esp32-s3-linha-a-01" }
    ProvisionResponse:
      type: object
      properties:
        tenant_id: { type: string, format: uuid }
        device_id: { type: string, format: uuid }
        device_label: { type: string }
        device_secret: { type: string }
        broker: { type: string, example: "mqtt.easysmart.com.br:1883" }
    BootstrapRequest:
      type: object
      required: [device_id, timestamp, signature]
      properties:
        device_id: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
        signature: { type: string }
    BootstrapResponse:
      type: object
      properties:
        status: { type: string, enum: [unclaimed, claimed, active, suspended, revoked] }
        device_id: { type: string, format: uuid }
        poll_interval: { type: integer, example: 60 }
    SecretRequest:
      type: object
      required: [device_id, timestamp, signature]
      properties:
        device_id: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
        signature: { type: string }
    SecretResponse:
      type: object
      properties:
        device_secret: { type: string }
        expires_at: { type: string, format: date-time }
    ResetRequest:
      type: object
      required: [device_id, confirmation]
      properties:
        device_id: { type: string, format: uuid }
        confirmation: { type: string, example: RESET }
    ResetResponse:
      type: object
      properties:
        status: { type: string, example: ok }
        message: { type: string, example: "Device reset to unclaimed" }
    DeviceSummary:
      type: object
      properties:
        device_id: { type: string, format: uuid }
        device_label: { type: string }
        status: { type: string, enum: [unclaimed, claimed, active, suspended, revoked] }
        firmware_version: { type: string, nullable: true }
        last_seen_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
    TelemetryWebhookRequest:
      type: object
      required: [topic, payload]
      properties:
        clientid: { type: string }
        topic: { type: string }
        payload:
          type: object
          example: { "value": 23.5 }
        timestamp: { type: string }
    TelemetryWebhookResponse:
      type: object
      properties:
        success: { type: boolean }
        device_id: { type: string, format: uuid }
        slot: { type: integer }
    LatestTelemetry:
      type: object
      properties:
        device_id: { type: string, format: uuid }
        slot: { type: integer }
        value: { type: object }
        timestamp: { type: string, format: date-time }
    ActiveSlotsResponse:
      type: object
      properties:
        device_id: { type: string, format: uuid }
        slots:
          type: array
          items: { type: integer }

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      responses:
        "200":
          description: OK
  /health/live:
    get:
      tags: [Health]
      summary: Liveness check
      responses:
        "200":
          description: OK
  /health/ready:
    get:
      tags: [Health]
      summary: Readiness check (Postgres + TimescaleDB + Redis)
      responses:
        "200":
          description: All dependencies OK
        "503":
          description: One or more dependencies unavailable
  /metrics:
    get:
      tags: [Health]
      summary: Prometheus metrics
      responses:
        "200":
          description: OK

  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register user (creates tenant automatically)
      description: |
        First user becomes super_admin. Subsequent users get their own tenant as tenant_admin.
        Password must be 8+ chars with uppercase, lowercase, number, and special char.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "400":
          description: Invalid input (email format, password strength)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: Email already registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Account not active
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh tokens (rotation with blacklist)
      description: |
        Issues new access + refresh tokens. Old refresh token is blacklisted (one-time use).
        Returns current role/tenant/permissions from DB (not from old token).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RefreshResponse" }
        "401":
          description: Invalid or already-used refresh token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Account not active
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "503":
          description: Token validation unavailable (Redis down, fail-closed)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /api/devices:
    get:
      tags: [Devices]
      summary: List devices for current tenant
      description: |
        Returns all devices belonging to the authenticated user's tenant.
        Requires JWT with `devices:read` permission. Results are scoped by tenant_id.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DeviceSummary" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Missing devices:read permission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /api/devices/claim:
    post:
      tags: [Devices]
      summary: Claim device (assign to tenant)
      description: |
        Validates claim_code against bcrypt hash. Generates device_secret (cached in Redis 5min for one-time retrieval).
        Requires JWT with `devices:provision` permission.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClaimRequest" }
      responses:
        "200":
          description: Device claimed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClaimResponse" }
        "401":
          description: Invalid claim code or unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Device not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: Device already claimed or missing claim code
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /api/devices/provision:
    post:
      tags: [Devices]
      summary: Direct device provisioning (authenticated user)
      description: |
        Creates a new claimed device for the current tenant and returns MQTT credentials.
        If `device_label` is omitted, backend generates one automatically.
        Requires JWT with `devices:provision` permission.
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProvisionRequest" }
      responses:
        "201":
          description: Device created and credentials issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProvisionResponse" }
        "400":
          description: Invalid payload
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Missing devices:provision permission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: device_label already exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /api/devices/bootstrap:
    post:
      tags: [Devices]
      summary: Device bootstrap (HMAC polling)
      description: |
        Device polls this endpoint to check its status. Signature is HMAC-SHA256(MANUFACTURING_MASTER_KEY, device_id:timestamp).
        Timestamp must be within configured skew window (default 5min).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BootstrapRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BootstrapResponse" }
        "401":
          description: Invalid HMAC signature or timestamp
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Device not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /api/devices/secret:
    post:
      tags: [Devices]
      summary: Device secret retrieval (HMAC, one-time)
      description: |
        Returns device_secret for MQTT authentication only once.
        Device must be in 'claimed' status and secret must still be available in cache.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SecretRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SecretResponse" }
        "401":
          description: Invalid HMAC signature or timestamp
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Device not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: Device not ready, secret already delivered, or secret unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "503":
          description: Cache unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /api/devices/reset:
    post:
      tags: [Devices]
      summary: Reset device to unclaimed
      description: |
        Resets device status, clears tenant/owner/secrets. Requires JWT with `devices:write` permission.
        tenant_admin/super_admin can reset any device in their tenant; tenant_user only their own.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResetRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResetResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Device not found or not authorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /api/telemetry:
    post:
      tags: [Telemetry]
      summary: Telemetry webhook (EMQX Rule Engine)
      description: |
        Receives telemetry from EMQX Rule Engine. Validates device, applies rate limiting,
        persists to TimescaleDB (with RLS), updates Redis cache, and marks device as active.
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TelemetryWebhookRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TelemetryWebhookResponse" }
        "400":
          description: Invalid JSON, topic format, or timestamp
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Invalid API key
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Device not found or inactive
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429":
          description: Rate limit exceeded (device or slot)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Internal server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "503":
          description: Rate limiter unavailable (fail-closed mode)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /api/telemetry/latest:
    get:
      tags: [Telemetry]
      summary: Latest telemetry from Redis cache
      description: |
        Returns the most recent cached value for a device/slot scoped to the authenticated tenant.
        Requires JWT with `telemetry:read`.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: device_id
          schema: { type: string, format: uuid }
          required: false
        - in: query
          name: device_label
          schema: { type: string }
          required: false
        - in: query
          name: slot
          schema: { type: integer }
          required: true
      responses:
        "200":
          description: OK (returns empty object {} if no cache)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LatestTelemetry" }
        "400":
          description: Missing slot or device identifier
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Missing telemetry:read permission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Device not found or inactive
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "503":
          description: Cache unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /api/telemetry/slots:
    get:
      tags: [Telemetry]
      summary: Active slots from Redis cache
      description: |
        Returns sorted list of slot numbers that have cached latest values,
        scoped to the authenticated tenant. Requires JWT with `telemetry:read`.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: device_id
          schema: { type: string, format: uuid }
          required: false
        - in: query
          name: device_label
          schema: { type: string }
          required: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActiveSlotsResponse" }
        "400":
          description: Missing device identifier
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Missing telemetry:read permission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Device not found or inactive
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "503":
          description: Cache unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }